chore: reorganize behaviors
