--- Behaviors/Nuclear/Classic/Druid.js
@@ -1,16 +1,0 @@
-import { Behavior, BehaviorContext } from '../../../Core/Behavior';
-import * as bt from '../../../Core/BehaviorTree';
-import Specialization from '../../../Enums/Specialization';
-import common from '../../../Core/Common';
-
-export class DruidRestoBehavior extends Behavior {
-  context = BehaviorContext.Any;
-  specialization = Specialization.Druid.Restoration;
-  version = wow.GameVersion.Classic;
-
-  build() {
-    return new bt.Selector(
-      common.waitForCastOrChannel(),
-    );
-  }
-}
