+++ Behaviors/Nuclear/WarriorProt.js
