+++ Behaviors/jmr/JmrSimcFury.js
