+++ Behaviors/neer/InitialHunter.js
