+++ Behaviors/soviet/DemonhunterHavoc.js
