+++ Behaviors/soviet/PaladinHolyPvP.js
