+++ Behaviors/soviet/PreservationPvP.js
@@ -8,7 +8,7 @@ import { DispelPriority } from "@/Data/Dispels";
 import { WoWDispelType } from "@/Enums/Auras";
 import { defaultHealTargeting as heal } from "@/Targeting/HealTargeting";
 import Settings from "@/Core/Settings";
-import EvokerCommon from "@/Behaviors/Nuclear/Retail/Evoker/EvokerCommon";
+import EvokerCommon from "@/Behaviors/EvokerCommon";
 
 const auras = {
   reversion: 366155,
