+++ Behaviors/soviet/PriestDisciplinePVE.js
